Create function getAllCharacters to follow SOLID design principle
The function getCharacterData was meant to return a single character or an array of all characters depending if ID was being passed or not.

To follow best practices in object-oriented software development I decided to split the function in two. This way each one gets a single responsibility, following SOLID design principles.

Now we have:
- getCharacterByID: Returns a character object from dynamic URL, depending on ID provided. Throws an error if no ID is provided.
- getAllCharacters: Returns an array of all character objects.